backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
sqlmodel>=0.0.14
supabase>=2.0.0
pydantic[email]>=2.0.0
python-dotenv>=1.0.0
requests>=2.31.0
psycopg2-binary>=2.9.9
python-multipart>=0.0.6
pytest>=8.0.0
pytest-cov>=7.0.0
httpx>=0.28.0
